+++ src/tasks/ReceiveTask.js
@@ -15,7 +15,7 @@ export function ReceiveTaskBehaviour(activity) {
   const reference = behaviour.messageRef || {name: 'anonymous'};
 
   const referenceElement = reference.id && activity.getActivityById(reference.id);
-  const loopCharacteristics = behaviour.loopCharacteristics && behaviour.loopCharacteristics.Behaviour(activity, behaviour.loopCharacteristics);
+  const loopCharacteristics = behaviour.loopCharacteristics && new behaviour.loopCharacteristics.Behaviour(activity, behaviour.loopCharacteristics);
 
   const source = {
     id,