+++ src/tasks/Task.js
@@ -6,7 +6,7 @@ export default function Task(activityDef, context) {
 
 export function TaskBehaviour(activity) {
   const {id, type, behaviour, broker} = activity;
-  const loopCharacteristics = behaviour.loopCharacteristics && behaviour.loopCharacteristics.Behaviour(activity, behaviour.loopCharacteristics);
+  const loopCharacteristics = behaviour.loopCharacteristics && new behaviour.loopCharacteristics.Behaviour(activity, behaviour.loopCharacteristics);
 
   const source = {
     id,