+++ test/tasks/LoopCharacteristics-test.js
@@ -2,6 +2,7 @@ import Environment from '../../src/Environment';
 import LoopCharacteristics from '../../src/tasks/LoopCharacteristics';
 import {ActivityBroker} from '../../src/EventBroker';
 import {Logger} from '../helpers/testHelpers';
+import {ActivityError} from '../../src/error/Errors';
 
 describe('LoopCharacteristics', () => {
   let task;
@@ -13,9 +14,9 @@ describe('LoopCharacteristics', () => {
     task.broker.bindQueue('execute-q', 'execution', '#');
   });
 
-  describe('ctorish(activity, loopCharacteristics)', () => {
+  describe('constructor(activity, loopCharacteristics)', () => {
     it('returns loop characteristics api with execute function', () => {
-      const loop = LoopCharacteristics(task, {
+      const loop = new LoopCharacteristics(task, {
         behaviour: {
           loopCardinality: 3,
           isSequential: true,
@@ -31,25 +32,27 @@ describe('LoopCharacteristics', () => {
       expect(loop).to.have.property('execute').that.is.a('function');
     });
 
-    it('returns NO loop characteristics if collection and cardinality is absent', () => {
+    it('returns loop characteristics if called without new', () => {
       const loop = LoopCharacteristics(task, {
         behaviour: {
+          loopCardinality: 3,
           isSequential: true,
+          collection: '${environment.variables.list}',
+          elementVariable: 'testitem',
         },
       });
 
-      expect(loop).to.not.be.ok;
-    });
-
-    it('empty behaviour is ok and returns no characteristics', () => {
-      const loop = LoopCharacteristics(task, {});
-      expect(loop).to.not.be.ok;
+      expect(loop).to.have.property('isSequential', true);
+      expect(loop).to.have.property('loopCardinality', 3);
+      expect(loop).to.have.property('collection', '${environment.variables.list}');
+      expect(loop).to.have.property('elementVariable', 'testitem');
+      expect(loop).to.have.property('execute').that.is.a('function');
     });
   });
 
   describe('execute(executeMessage)', () => {
     it('throws if executeMessage is missing', () => {
-      const loop = LoopCharacteristics(task, {
+      const loop = new LoopCharacteristics(task, {
         behaviour: {
           loopCardinality: 3,
           isSequential: true,
@@ -59,46 +62,22 @@ describe('LoopCharacteristics', () => {
       expect(loop.execute).to.throw(/requires message/);
     });
 
-    it('publishes error on activity broker if loopCardinality is not a number', (done) => {
-      const loop = LoopCharacteristics(task, {
+    it('throws error if loopCardinality is not a number', () => {
+      const loop = new LoopCharacteristics(task, {
         behaviour: {
           loopCardinality: '3 pcs',
           isSequential: true,
         },
       });
 
-      task.once('error', (err) => {
-        expect(err).to.match(/NaN/i);
-        done();
-      });
-
-      loop.execute({
-        content: {
-          isRootScope: true,
-          executionId: 'parent-execution-id',
-        },
-      });
-    });
-
-    it('publishes error on activity broker if loopCardinality is not a number', (done) => {
-      const loop = LoopCharacteristics(task, {
-        behaviour: {
-          loopCardinality: '3 pcs',
-          isSequential: true,
-        },
-      });
-
-      task.once('error', (err) => {
-        expect(err).to.match(/NaN/i);
-        done();
-      });
-
-      loop.execute({
-        content: {
-          isRootScope: true,
-          executionId: 'parent-execution-id',
-        },
-      });
+      expect(() => {
+        loop.execute({
+          content: {
+            isRootScope: true,
+            executionId: 'parent-execution-id',
+          },
+        });
+      }).to.throw(ActivityError, /cardinality/i);
     });
   });
 
@@ -109,7 +88,7 @@ describe('LoopCharacteristics', () => {
         messages.push(msg);
       }, {noAck: true});
 
-      const loop = LoopCharacteristics(task, {
+      const loop = new LoopCharacteristics(task, {
         behaviour: {
           loopCardinality: 3,
           isSequential: true,
@@ -117,6 +96,7 @@ describe('LoopCharacteristics', () => {
       });
 
       loop.execute({
+        fields: {},
         content: {
           isRootScope: true,
           executionId: 'parent-execution-id',
@@ -144,7 +124,7 @@ describe('LoopCharacteristics', () => {
         messages.push(msg);
       }, {noAck: true});
 
-      const loop = LoopCharacteristics(task, {
+      const loop = new LoopCharacteristics(task, {
         behaviour: {
           loopCardinality: 3,
           isSequential: true,
@@ -152,6 +132,7 @@ describe('LoopCharacteristics', () => {
       });
 
       loop.execute({
+        fields: {},
         content: {
           isRootScope: true,
           executionId: 'parent-execution-id',
@@ -170,7 +151,7 @@ describe('LoopCharacteristics', () => {
         messages.push(msg);
       }, {noAck: true});
 
-      const loop = LoopCharacteristics(task, {
+      const loop = new LoopCharacteristics(task, {
         behaviour: {
           isSequential: true,
           loopCardinality: 3,
@@ -178,6 +159,7 @@ describe('LoopCharacteristics', () => {
       });
 
       loop.execute({
+        fields: {},
         content: {
           id: 'task',
           type: 'bpmn:Task',
@@ -215,7 +197,7 @@ describe('LoopCharacteristics', () => {
         messages.push(msg);
       }, {noAck: true});
 
-      const loop = LoopCharacteristics(task, {
+      const loop = new LoopCharacteristics(task, {
         behaviour: {
           isSequential: true,
           loopCardinality: 3,
@@ -223,6 +205,7 @@ describe('LoopCharacteristics', () => {
       });
 
       loop.execute({
+        fields: {},
         content: {
           isRootScope: true,
           executionId: 'parent-execution-id',
@@ -262,7 +245,7 @@ describe('LoopCharacteristics', () => {
         messages.push(msg);
       }, {noAck: true});
 
-      const loop = LoopCharacteristics(task, {
+      const loop = new LoopCharacteristics(task, {
         behaviour: {
           isSequential: true,
           loopCardinality: 3,
@@ -270,6 +253,7 @@ describe('LoopCharacteristics', () => {
       });
 
       loop.execute({
+        fields: {},
         content: {
           isRootScope: true,
           executionId: 'parent-execution-id',
@@ -303,7 +287,7 @@ describe('LoopCharacteristics', () => {
         task.broker.publish('execution', 'execute.completed', {...msg.content});
       }, {noAck: true});
 
-      const loop = LoopCharacteristics(task, {
+      const loop = new LoopCharacteristics(task, {
         behaviour: {
           isSequential: true,
           loopCardinality: 3,
@@ -311,6 +295,7 @@ describe('LoopCharacteristics', () => {
       });
 
       loop.execute({
+        fields: {},
         content: {
           isRootScope: true,
           executionId: 'parent-execution-id',
@@ -340,7 +325,7 @@ describe('LoopCharacteristics', () => {
         }
       }, {noAck: true, consumerTag: 'completed-consumer'});
 
-      const loop = LoopCharacteristics(task, {
+      const loop = new LoopCharacteristics(task, {
         behaviour: {
           isSequential: true,
           loopCardinality: 3,
@@ -348,6 +333,7 @@ describe('LoopCharacteristics', () => {
       });
 
       loop.execute({
+        fields: {},
         content: {
           isRootScope: true,
           executionId: 'parent-execution-id',
@@ -370,7 +356,7 @@ describe('LoopCharacteristics', () => {
 
       task.environment.variables.items = ['item 1', 'item 2', 'item 3', 'item 4'];
 
-      const loop = LoopCharacteristics(task, {
+      const loop = new LoopCharacteristics(task, {
         behaviour: {
           isSequential: true,
           loopCardinality: 10,
@@ -379,6 +365,7 @@ describe('LoopCharacteristics', () => {
       });
 
       loop.execute({
+        fields: {},
         content: {
           isRootScope: true,
           executionId: 'parent-execution-id',
@@ -417,7 +404,7 @@ describe('LoopCharacteristics', () => {
         }
       }, {noAck: true});
 
-      const loop = LoopCharacteristics(task, {
+      const loop = new LoopCharacteristics(task, {
         behaviour: {
           isSequential: true,
           loopCardinality: 3,
@@ -426,6 +413,7 @@ describe('LoopCharacteristics', () => {
       });
 
       loop.execute({
+        fields: {},
         content: {
           isRootScope: true,
           executionId: 'parent-execution-id',
@@ -446,7 +434,7 @@ describe('LoopCharacteristics', () => {
     it('root api stop message drops consumers', () => {
       task.environment.variables.items = ['item 1', 'item 2', 'item 3', 'item 4'];
 
-      const loop = LoopCharacteristics(task, {
+      const loop = new LoopCharacteristics(task, {
         behaviour: {
           isSequential: true,
           collection: '${environment.variables.items}',
@@ -454,6 +442,7 @@ describe('LoopCharacteristics', () => {
       });
 
       loop.execute({
+        fields: {},
         content: {
           isRootScope: true,
           executionId: 'parent-execution-id',
@@ -472,7 +461,7 @@ describe('LoopCharacteristics', () => {
           messages.push(msg);
         }, {noAck: true});
 
-        const loop = LoopCharacteristics(task, {
+        const loop = new LoopCharacteristics(task, {
           behaviour: {
             isSequential: true,
             loopCardinality: 3,
@@ -511,7 +500,7 @@ describe('LoopCharacteristics', () => {
           messages.push(msg);
         }, {noAck: true});
 
-        const loop = LoopCharacteristics(task, {
+        const loop = new LoopCharacteristics(task, {
           behaviour: {
             isSequential: true,
             loopCardinality: 3,
@@ -549,7 +538,7 @@ describe('LoopCharacteristics', () => {
           messages.push(msg);
         }, {noAck: true});
 
-        const loop = LoopCharacteristics(task, {
+        const loop = new LoopCharacteristics(task, {
           behaviour: {
             isSequential: true,
             loopCardinality: 3,
@@ -585,7 +574,7 @@ describe('LoopCharacteristics', () => {
           messages.push(msg);
         }, {noAck: true});
 
-        const loop = LoopCharacteristics(task, {
+        const loop = new LoopCharacteristics(task, {
           behaviour: {
             isSequential: true,
             loopCardinality: 3,
@@ -645,13 +634,14 @@ describe('LoopCharacteristics', () => {
         messages.push(msg);
       }, {noAck: true});
 
-      const loop = LoopCharacteristics(task, {
+      const loop = new LoopCharacteristics(task, {
         behaviour: {
           loopCardinality: 3,
         },
       });
 
       loop.execute({
+        fields: {},
         content: {
           id: 'task',
           type: 'bpmn:Task',
@@ -734,13 +724,14 @@ describe('LoopCharacteristics', () => {
 
       task.environment.variables.items = ['item 1', 'item 2', 'item 3', 'item 4'];
 
-      const loop = LoopCharacteristics(task, {
+      const loop = new LoopCharacteristics(task, {
         behaviour: {
           collection: '${environment.variables.items}',
         },
       });
 
       loop.execute({
+        fields: {},
         content: {
           isRootScope: true,
           executionId: 'parent-execution-id',
@@ -776,13 +767,14 @@ describe('LoopCharacteristics', () => {
         messages.push(msg);
       }, {noAck: true});
 
-      const loop = LoopCharacteristics(task, {
+      const loop = new LoopCharacteristics(task, {
         behaviour: {
           loopCardinality: 3,
         },
       });
 
       loop.execute({
+        fields: {},
         content: {
           id: 'task',
           type: 'bpmn:Task',
@@ -808,13 +800,14 @@ describe('LoopCharacteristics', () => {
         messages.push(msg);
       }, {noAck: true});
 
-      const loop = LoopCharacteristics(task, {
+      const loop = new LoopCharacteristics(task, {
         behaviour: {
           loopCardinality: 3,
         },
       });
 
       loop.execute({
+        fields: {},
         content: {
           isRootScope: true,
           executionId: 'parent-execution-id',
@@ -845,7 +838,7 @@ describe('LoopCharacteristics', () => {
         startMessages.push(msg);
       }, {noAck: true});
 
-      const loop = LoopCharacteristics(task, {
+      const loop = new LoopCharacteristics(task, {
         behaviour: {
           loopCardinality: 3,
           completionCondition: '${content.output.stopLoop}',
@@ -853,6 +846,7 @@ describe('LoopCharacteristics', () => {
       });
 
       loop.execute({
+        fields: {},
         content: {
           isRootScope: true,
           executionId: 'parent-execution-id',
@@ -873,13 +867,14 @@ describe('LoopCharacteristics', () => {
     it('root api stop message drops consumers', () => {
       task.environment.variables.items = ['item 1', 'item 2', 'item 3', 'item 4'];
 
-      const loop = LoopCharacteristics(task, {
+      const loop = new LoopCharacteristics(task, {
         behaviour: {
           collection: '${environment.variables.items}',
         },
       });
 
       loop.execute({
+        fields: {},
         content: {
           isRootScope: true,
           executionId: 'parent-execution-id',
@@ -898,7 +893,7 @@ describe('LoopCharacteristics', () => {
           messages.push(msg);
         }, {noAck: true});
 
-        const loop = LoopCharacteristics(task, {
+        const loop = new LoopCharacteristics(task, {
           behaviour: {
             loopCardinality: 3,
             completionCondition: '${content.output.stopLoop}',
@@ -926,7 +921,7 @@ describe('LoopCharacteristics', () => {
           messages.push(msg);
         }, {noAck: true});
 
-        const loop = LoopCharacteristics(task, {
+        const loop = new LoopCharacteristics(task, {
           behaviour: {
             loopCardinality: 3,
             completionCondition: '${content.output.stopLoop}',
@@ -955,7 +950,7 @@ describe('LoopCharacteristics', () => {
           messages.push(msg);
         }, {noAck: true});
 
-        const loop = LoopCharacteristics(task, {
+        const loop = new LoopCharacteristics(task, {
           behaviour: {
             loopCardinality: 3,
           },